Share remote file browser; show SSH connection names
- Extract RemoteFileBrowser from SshTestDialog so it can be reused (adds a directory pick mode; SshTestDialog now delegates to it).
- Profiler launcher: enable the Target Browse buttons in Remote (SSH) mode to open the shared browser; local mode unchanged.
- Add SshConnectionConfig::DisplayLabel() and use it in the connection chips (name if set, else user@host:port).
